LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/ordering/parser-blocking-expected.txt

CONSOLE MESSAGE: Not executing script because it moved between documents during fetching


PASS Script elements (parser-blocking) still block the parser in the original Document after moved to another Document
